// contracts/generated/typescript/models/PageProtection.ts
export interface PageProtection {
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    pageId?: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    pageTitle?: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    namespaceId?: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    namespaceKey?: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    normalizedTitle?: string;
    /**
     *
     * @type {PageProtectionActionTypeEnum}
     * @memberof PageProtection
     */
    actionType: PageProtectionActionTypeEnum;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    requiredRoleId: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    requiredRoleKey: string;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    requiredRoleName: string;
    /**
     *
     * @type {Date}
     * @memberof PageProtection
     */
    expiresAt?: Date;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    createdBy: string;
    /**
     *
     * @type {Date}
     * @memberof PageProtection
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof PageProtection
     */
    revokedAt?: Date;
    /**
     *
     * @type {string}
     * @memberof PageProtection
     */
    revokedBy?: string;
}


/**
 * @export
 */
export const PageProtectionActionTypeEnum = {
    Create: 'create',
    Edit: 'edit',
    Rename: 'rename',
    Review: 'review',
    Apply: 'apply',
    BatchRollback: 'batch_rollback'
} as const;
export type PageProtectionActionTypeEnum = typeof PageProtectionActionTypeEnum[keyof typeof PageProtectionActionTypeEnum];